refactor: Report 도메인 서비스 계층 추상화
